Hide step + fields for Brazil
