+++ .env
@@ -1,6 +1,7 @@
 NODE_ENV=development
 
 REACT_APP_API=https://ti7vvvk450.execute-api.us-east-2.amazonaws.com/prod
+# REACT_APP_API=https://ozuva4g117.execute-api.us-east-2.amazonaws.com/dev
 
 REACT_APP_RECAPTCHA_KEY=6LcRhdIZAAAAAOI8zafkWF0N5SYOUjMszuEUnrJT
