+++ .gitignore
@@ -31,3 +31,4 @@ _redirects
 
 # Local Netlify folder
 .netlify
+deploy.sh
