+++ src/components/Recaptcha/style.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const RecaptchaContainer = styled.div`
+  margin: 20px 0;
+  
+  div {
+    margin: auto;
+  }
+`;
