+++ src/containers/SubmitSteps/Questionary/Step2/Step2.tsx
@@ -100,7 +100,7 @@ const Step2 = ({
 
   return (
     <>
-      <ProgressIndicator currentStep={3} totalSteps={4} />
+      <ProgressIndicator currentStep={metadata?.progressCurrent || 3} totalSteps={metadata?.progressTotal || 4} />
 
       <GrayExtraInfo>{t('questionary:caption')}</GrayExtraInfo>
