+++ src/containers/SubmitSteps/Questionary/Step3/Step3.tsx
@@ -90,7 +90,7 @@ const Step3 = ({
 
   return (
     <>
-      <ProgressIndicator currentStep={3} totalSteps={4} />
+      <ProgressIndicator currentStep={metadata?.progressCurrent || 3} totalSteps={metadata?.progressTotal || 4} />
 
       <GrayExtraInfo>{t('questionary:caption')}</GrayExtraInfo>
