+++ src/containers/SubmitSteps/Submission/ThankYou/ThankYou.tsx
@@ -4,13 +4,13 @@ import { useTranslation, Trans } from 'react-i18next';
 import { useStateMachine } from 'little-state-machine';
 import usePortal from 'react-useportal';
 
-// Header Control
-import useHeaderContext from 'hooks/useHeaderContext';
-
 // Components
 import SocialIcons from 'components/SocialIcons';
 import WizardButtons from 'components/WizardButtons';
 import Link from 'components/LinkGreen';
+
+// Utils
+import { resetStore } from 'utils/wizard';
 
 // Utils
 import { scrollToTop } from 'helper/scrollHelper';
@@ -19,6 +19,7 @@ import { scrollToTop } from 'helper/scrollHelper';
 import { feedbackForm } from 'data/feedbackForm';
 
 // Hooks
+import useHeaderContext from 'hooks/useHeaderContext';
 import usePWAHelpers from 'hooks/usePWAHelpers';
 
 import {
@@ -39,7 +40,7 @@ const ThankYou = (p: Wizard.StepProps) => {
 
   const [, setActiveStep] = useState(true);
   const { setDoGoBack, setTitle } = useHeaderContext();
-  const { state } = useStateMachine();
+  const { state, action } = useStateMachine(resetStore());
   const { handlePrompt, isInstalled, setIsInstalled } = usePWAHelpers(installPwaButtonId);
 
   const lang: FeedbackLanguage = state.welcome.language;
@@ -49,10 +50,11 @@ const ThankYou = (p: Wizard.StepProps) => {
   const submissionId = location.state?.submissionId;
 
   const handleNext = React.useCallback(() => {
+    action({});
     if (p.nextStep) {
       history.push(p.nextStep);
     }
-  }, [history, p.nextStep]);
+  }, [action, history, p.nextStep]);
 
   const handleDoBack = useCallback(() => {
     if (p.previousStep) {
