+++ src/hooks/useAxios.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-const client = axios.create({
+export const client = axios.create({
   baseURL: process.env.REACT_APP_API,
 });
